requirements.txt
alive-progress==2.4.1
numpy==1.22.4
Pillow==9.2.0
